+++ app/core/Router/PrefixRouter.php
@@ -7,18 +7,18 @@ use PhpSPA\Http\Router;
 
 trait PrefixRouter {
    /**
-   * The base URI of the application.
-   * This is used to determine the root path for routing and resource loading.
-   *
-   * @var string
-   */
+    * The base URI of the application.
+    * This is used to determine the root path for routing and resource loading.
+    *
+    * @var string
+    */
    public static string $request_uri;
 
    /**
-   * Indicates whether the application should treat string comparisons as case sensitive.
-   *
-   * @var bool Defaults to false, meaning string comparisons are case insensitive by default.
-   */
+    * Indicates whether the application should treat string comparisons as case sensitive.
+    *
+    * @var bool Defaults to false, meaning string comparisons are case insensitive by default.
+    */
    protected bool $defaultCaseSensitive = false;
 
    /**
@@ -30,12 +30,13 @@ trait PrefixRouter {
    protected array $static = [];
 
    /**
-   * @param array{
-   *   path: string,
-   *   handler: callable
-   * } $prefix
-   */
-   protected function handlePrefix(array $prefix, array $middlewares = [], bool $return = false) {
+    * @param array{
+    *   path: string,
+    *   handler: callable
+    * } $prefix
+    */
+   protected function handlePrefix (array $prefix, array $middlewares = [], bool $return = false)
+   {
       // --- Replacing first and last forward slashes, $request_uri will be empty if req uri is / ---
       $request_uri = trim(static::$request_uri, '/');
       $prefixPath = trim($prefix['path'], '/');
@@ -52,7 +53,7 @@ trait PrefixRouter {
             prefix: $prefix['path'],
             caseSensitive: $this->defaultCaseSensitive,
             middlewares: $middlewares,
-            return: $return
+            return: $return,
          );
 
          if ($return) {
@@ -71,7 +72,8 @@ trait PrefixRouter {
       }
    }
 
-   protected function resolveStaticPath(bool $return) {
+   protected function resolveStaticPath (bool $return)
+   {
       $request_uri = trim(static::$request_uri, '/');
 
       foreach ($this->static as $static) {
@@ -81,13 +83,13 @@ trait PrefixRouter {
          if ($path === $route) {
             $staticRoute = substr($request_uri, \strlen($route));
             $filePath = rtrim($static['staticPath'], '/') . '/' . ltrim($staticRoute, '/');
-            
+
             if (is_file($filePath)) {
                if ($return) {
-                  return (new Response())->sendFile($filePath)->__toString();
+                  return new Response()->sendFile($filePath)->__toString();
                }
 
-               (new Response())->sendFile($filePath)->send();
+               new Response()->sendFile($filePath)->send();
             }
          }
       }