+++ lib/types/RuntimeInterfaces.ts
@@ -18,6 +18,11 @@ export interface EventPayload {
    data?: string
 }
 
+export interface RuntimeConfig {
+   preserveUpdatedHtmlState: boolean,
+   waitForStyles: boolean
+}
+
 export type CurrentRoutesObject = Record<string, {
    route: URL,
    exact: boolean,